✨ package.jsonのテストスクリプトを更新し、vitestのプロジェクト指定を追加しました。また、vitest.config.tsを削除し、vitest.workspace.tsにユニットテストとStorybookテストの設定を統合しました。GitHub Actionsのワークフローにgitの設定を追加しました。
