+++ package.json
@@ -13,8 +13,8 @@
     "start": "pnpm run '/^start:.*$/'",
     "start:next": "serve out -p 8000",
     "start:storybook": "serve storybook-static -p 6008",
-    "test:unit": "vitest",
-    "test:unit-coverage": "vitest run --coverage",
+    "test:unit": "vitest --project=unit",
+    "test:unit-coverage": "pnpm test:unit run --coverage",
     "test:unit-ui": "vitest --reporter='html' --outputFile='.vitest/ui/index.html' --ui",
     "test:storybook": "vitest --project=storybook",
     "eslint:check": "next lint",
